Adds --search and -s flags
This update adds a search feature,
Which allows you to type in a partial Module name or Function name,
And find all the modules or functions that match.

This combines `:code.all_loaded()`, which returns all modules,
And `IEx.Autocomplete.exports`,
which grabs the exported functions and macros from each module.

This implementation currently doesn't support arity in search,
And uses `String.starts_with?` for filtering,
I chose `starts_with?` over `contains?`,
Because the latter results in partial matches that may be unrelated to your query.