Adds "completed" Search module
This commit completes work on Search for now.
It implements all three cases that I had before,
With the exception that bare function search can't find Kernel functions.
This is because IEx.Introspection.decompose returns {Kernel, atom} whenever it can't find that atom as a function in any other module.

There are two paths forward here:
1. We create a subfunction that prevents {Kernel, fun} from creating an infinite recursion on itself,
2. We vendor IEx.Introspection.decompose and modify it to return a :not_found atom when a function isn't found.

Theoretically, option 2 would allow us to still find Kernel functions while creating a case for atoms not found elsewhere.

It would involve determining how decompose performs its work,
And changing the fall-back case when a function can't be found.
It remains to be seen if Kernel functions can still be found without the fallback.