Completes Helpers interface module
This commit completes existing work on the Helpers module,
Which is designed to act as an interface between Exhelp
And existing internal functions from IEx and Elixir.

This has already proven useful in the decision to vendor `decompose`,
Since I only needed to change the interface and everything else continued to work.
This helps protect against the possibility that the IEx helper functions change in a way that breaks Exhelp in the future.